Replace doubles with floats

+++ ray.h
@@ -21,22 +21,22 @@ class ray {
             : orig(origin), dir(direction), tm(0)
         {}
 
-        ray(const point3& origin, const vec3& direction, double time)
+        ray(const point3& origin, const vec3& direction, float time)
             : orig(origin), dir(direction), tm(time)
         {}
 
         point3 origin() const  { return orig; }
         vec3 direction() const { return dir; }
-        double time() const    { return tm; }
+        float time() const    { return tm; }
 
-        point3 at(double t) const {
+        point3 at(float t) const {
             return orig + t*dir;
         }
 
     public:
         point3 orig;
         vec3 dir;
-        double tm;
+        float tm;
 };
 
 #endif
